chore: generify charting solution
