+++ src/charts/chart-data.model.ts
@@ -1,0 +1,9 @@
+import { attribute } from '@aws/dynamodb-data-mapper-annotations';
+
+export class ChartData {
+  @attribute()
+  id!: string;
+
+  @attribute()
+  timestamp!: number;
+}
