+++ src/config/constants.ts
@@ -1,14 +1,11 @@
 import { SwaggerSettings } from '@tsed/swagger';
 import { Stage } from './enums/stage.enum';
-import { getEnvVar } from './util';
+import { getEnvVar } from './config.utils';
 
 // time constants
 export const ONE_MINUTE_SECONDS = 60;
-export const ONE_MINUTE_MS = ONE_MINUTE_SECONDS * 1000;
 export const ONE_DAY_SECONDS = ONE_MINUTE_SECONDS * 60 * 24;
-export const ONE_DAY_MS = ONE_DAY_SECONDS * 1000;
 export const ONE_YEAR_SECONDS = ONE_DAY_SECONDS * 365;
-export const ONE_YEAR_MS = ONE_YEAR_SECONDS * 1000;
 
 // data access constants
 export const APY_SNAPSHOTS_DATA = getEnvVar('APY_SNAPSHOTS_DATA');
